multi-agent-voice-system/requirements.txt
# Core framework
langgraph>=0.2.0
langchain>=0.3.0

# LiveKit voice agents
livekit-agents[openai,silero,deepgram,cartesia]~=1.0

# LLM providers
cerebras-cloud-sdk
openai

# Vector store and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# PyTorch for intent classification
torch>=2.0.0
transformers>=4.30.0

# Utilities
numpy>=1.24.0
python-dotenv>=1.0.0

# Web scraping and data processing
beautifulsoup4
requests
pdfplumber

# Development tools
pytest>=7.0.0
black
pylint
